requirements.txt
cairocffi==1.2.0
cffi==1.14.4
numpy==1.19.4
pandas==1.1.4
pycairo==1.20.0
pycparser==2.20
python-dateutil==2.8.1
python-igraph==0.8.3
pytz==2020.4
six==1.15.0
texttable==1.6.3
